requirements.txt
#pip3 install -r requirements.txt

Werkzeug>=0.14.1
Flask>=1.0.2
Jinja2>=2.10
gunicorn>=19.9.0
#Flask-Login>=0.4.1
psycopg2-binary>=2.7.0

#[requires]
#python_version > 3.5
